Move external resource analyzers & transfomers into separate folers
